requirements.txt
boto3==1.34.69
botocore==1.34.69
eventlet==0.35.2
Flask==3.0.2
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio>=4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.1
awscli-local==0.20 